[faet] 현재 유저와 작성자가 같은 게시글만 삭제 할 수 있도록 코드 생성

+++ pages/api/post/delete.js
@@ -1,12 +1,24 @@
 import { connectDB } from "@/util/database";
 import { ObjectId } from "mongodb";
+import { getServerSession } from "next-auth";
+import { authOptions } from "../auth/[...nextauth]";
 
 export default async function handler(요청, 응답) {
   if (요청.method == "POST") {
-    let db = (await connectDB).db("forum");
-    let result = await db
+    let session = await getServerSession(요청, 응답, authOptions);
+
+    const db = (await connectDB).db("forum");
+    let 찾은거 = await db
       .collection("post")
-      .deleteOne({ _id: new ObjectId(요청.body) });
-    응답.status(200).json("삭제완료");
+      .findeOne({ _id: new ObjectId(요청.body) });
+
+    if (찾은거.author == session.user.email) {
+      let result = await db
+        .collection("post")
+        .deleteOne({ _id: new ObjectId(요청.body) });
+      return 응답.status(200).json("삭제완료");
+    } else {
+      return 응답.status(500).json("현재유저와 작성자 불일치");
+    }
   }
 }
